rewrite mypets/league cmd to ts
